Routes, switch, links, and dashboard edits
