+++ package.json
@@ -7,9 +7,7 @@
     "@aws-amplify/interactions": "^3.3.29",
     "@aws-amplify/storage": "^3.3.29",
     "@aws-amplify/ui": "^2.0.2",
-
     "@aws-amplify/ui-react": "^1.0.7",
-
     "@aws-amplify/xr": "^2.2.29",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
@@ -17,17 +15,14 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "aws-amplify": "^3.3.26",
-
     "mdb-react-ui-kit": "^1.0.0-beta3",
-
     "react": "^17.0.2",
     "react-clock": "^3.0.0",
     "react-countdown-circle-timer": "^2.5.1",
     "react-dom": "^17.0.2",
-
     "react-router-dom": "^5.2.0",
-
     "react-scripts": "4.0.3",
+    "react-stopwatch": "^2.0.4",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
